Move ConstraintConversionService to api module of nrich-notification
