+++ nrich-notification-spring-boot-starter/src/main/java/net/croz/nrich/notification/starter/configuration/NrichNotificationAutoConfiguration.java
@@ -1,11 +1,11 @@
 package net.croz.nrich.notification.starter.configuration;
 
 import lombok.RequiredArgsConstructor;
+import net.croz.nrich.notification.api.service.ConstraintConversionService;
 import net.croz.nrich.notification.api.service.NotificationMessageResolverService;
 import net.croz.nrich.notification.api.service.NotificationResolverService;
 import net.croz.nrich.notification.api.service.BaseNotificationResponseService;
 import net.croz.nrich.notification.api.service.NotificationResponseService;
-import net.croz.nrich.notification.service.ConstraintConversionService;
 import net.croz.nrich.notification.service.DefaultConstraintConversionService;
 import net.croz.nrich.notification.service.DefaultNotificationResolverService;
 import net.croz.nrich.notification.service.MessageSourceNotificationMessageResolverService;
