+++ nrich-webmvc/src/test/java/net/croz/nrich/webmvc/WebmvcTestConfiguration.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import net.croz.nrich.logging.api.service.LoggingService;
 import net.croz.nrich.logging.service.Slf4jLoggingService;
 import net.croz.nrich.notification.api.service.BaseNotificationResponseService;
+import net.croz.nrich.notification.api.service.ConstraintConversionService;
 import net.croz.nrich.notification.api.service.NotificationMessageResolverService;
 import net.croz.nrich.notification.api.service.NotificationResolverService;
 import net.croz.nrich.notification.api.service.NotificationResponseService;
-import net.croz.nrich.notification.service.ConstraintConversionService;
 import net.croz.nrich.notification.service.DefaultConstraintConversionService;
 import net.croz.nrich.notification.service.DefaultNotificationResolverService;
 import net.croz.nrich.notification.service.MessageSourceNotificationMessageResolverService;
